refactor(cdk): Remove unnecessary `ecs` directory
There is more than ECS related code in this directory.
Therefore the naming doesn't make sense.
Remove it to simplify things.
